configure: changed rebuild -> fbuild
This is to avoid confusion, i.e. "I did a rebuild but now I'm still in
my home dir". Yes, because the old rebuild alias does a popd. When I say
"do a rebuild", that means re-run "build". "fbuild" is provided for
running a quick rebuild without having to popd afterwards.